+++ src/main/java/com/chromosundrift/vectorbrat/physics/TeleportBeamPhysics.java
@@ -7,7 +7,7 @@ package com.chromosundrift.vectorbrat.physics;
 public class TeleportBeamPhysics implements BeamPhysics {
 
     @Override
-    public void timeStep(float x, float y, float r, float g, float b, BeamState state, long nsTimeStep) {
+    public void timeStep(double x, double y, float r, float g, float b, BeamState state, long nsTimeStep) {
         state.xPos = x;
         state.yPos = y;
     }
